# src/swagger/swagger.json
{"swagger":"2.0","info":{"title":"DPE Service","version":"1.0.0","description":""},"tags":[{"name":"Group"},{"name":"Patient"},{"name":"SensoryBlock"}],"paths":{"/dpe/group/getGroups":{"get":{"tags":["Group"],"summary":"get groups","description":"get all patient groups","consumes":["application/json"],"produces":["application/json"],"parameters":[]}},"/dpe/patient/getPatients":{"get":{"tags":["Patient"],"summary":"get patients","description":"get all patients","consumes":["application/json"],"produces":["application/json"],"parameters":[]}},"/dpe/patient/addPatient":{"post":{"tags":["Patient"],"summary":"add new patient","description":"add new patient","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/patients"}}]}},"/dpe/patient/updateAnalgesiaData":{"post":{"tags":["Patient"],"summary":"update analgesia data","description":"update analgesia data to the patient","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"body","schema":{"$ref":"#/definitions/analgesia_data"}}]}},"/dpe/sensoryBlock/getSensoryBlocks":{"get":{"tags":["SensoryBlock"],"summary":"get sensory blocks","description":"get all sensory blocks","consumes":["application/json"],"produces":["application/json"],"parameters":[]}}},"definitions":{"analgesia_data":{"type":"object","properties":{"id":{"type":"integer"},"patientId":{"type":"string"},"timePoint":{"type":"integer"},"hasContraction":{"type":"boolean"},"vasScore":{"type":"integer"},"thoracicSensoryBlockLeft":{"type":"integer"},"thoracicSensoryBlockRight":{"type":"integer"},"sacralSensoryBlockLeft":{"type":"integer"},"sacralSensoryBlockRight":{"type":"integer"},"bromageScore":{"type":"integer"},"systolicBloodPressure":{"type":"integer"},"diastolicBloodPressure":{"type":"integer"},"heartRate":{"type":"integer"},"pulseOxygenSaturation":{"type":"float"},"ctime":{"type":"datetime"},"utime":{"type":"datetime"},"dtime":{"type":"datetime"}},"xml":{"name":"analgesia_data"}},"dpe_group":{"type":"object","properties":{"id":{"type":"integer"},"groupName":{"type":"string"}},"xml":{"name":"dpe_group"}},"observal_data":{"type":"object","properties":{"id":{"type":"integer"},"patientId":{"type":"string"},"hasCarbetocin":{"type":"boolean"},"hasHemabate":{"type":"boolean"},"localAnestheticConsumption":{"type":"integer"},"pcaCount":{"type":"integer"},"manualBolusCount":{"type":"integer"},"firstPcaTime":{"type":"datetime"},"firstManualBolusTime":{"type":"datetime"},"durationOfSecondStageOfLabor":{"type":"integer"},"hasEpiduralCatheterAdjuestment":{"type":"boolean"},"hasEpiduralCatheterReplacement":{"type":"boolean"},"hasPrenatalFever":{"type":"boolean"},"hasVasoactiveAgent":{"type":"boolean"},"isUnabled_ToPunctureDura":{"type":"boolean"},"hasNausea":{"type":"boolean"},"hasVomit":{"type":"boolean"},"hasPruritus":{"type":"boolean"},"hasHypotension":{"type":"boolean"},"hasCaesareanSection":{"type":"boolean"},"hasInstrumental":{"type":"boolean"},"hasPostduralPunctureHeadache":{"type":"boolean"},"hasBackPain":{"type":"boolean"},"hasParesthesia":{"type":"boolean"},"durationOfLaborAnalgesia":{"type":"integer"},"patientSatisfactionScore":{"type":"integer"},"hasAccidentalDuralPunture":{"type":"boolean"},"lateralEpisiotomyVasScore":{"type":"integer"},"hasLateralEpisiotomy":{"type":"boolean"},"birthTime":{"type":"datetime"},"foetalHeight":{"type":"integer"},"foetalWeight":{"type":"integer"},"oneMinuteApgarScore":{"type":"integer"},"fiveMinuteApgarScore":{"type":"integer"},"hasNicu":{"type":"boolean"},"nicuReason":{"type":"string"},"arterialPh":{"type":"float"},"arterialBe":{"type":"float"},"venousPh":{"type":"float"},"venousBe":{"type":"float"},"description":{"type":"string"},"ctime":{"type":"datetime"},"utime":{"type":"datetime"},"dtime":{"type":"datetime"}},"xml":{"name":"observal_data"}},"patients":{"type":"object","properties":{"id":{"type":"string"},"groupId":{"type":"integer"},"patientName":{"type":"string"},"age":{"type":"integer"},"gestationalDays":{"type":"integer"},"height":{"type":"float"},"weight":{"type":"float"},"initialVasScore":{"type":"integer"},"cervicalDilationAtTimeOfEa":{"type":"integer"},"heartRate":{"type":"integer"},"systolicBloodPressure":{"type":"integer"},"diastolicBloodPressure":{"type":"integer"},"foetalHeartRate":{"type":"integer"},"description":{"type":"string"},"ctime":{"type":"datetime"},"utime":{"type":"datetime"},"dtime":{"type":"datetime"}},"xml":{"name":"patients"}},"sensory_block":{"type":"object","properties":{"id":{"type":"integer"},"sensoryBlockName":{"type":"string"},"sensoryBlockValue":{"type":"integer"}},"xml":{"name":"sensory_block"}}}}